+++ extralit-frontend/v1/domain/entities/import/ImportAnalysis.ts
@@ -6,24 +6,6 @@
 // Basic field types supported in dataframes
 export type FieldType = "string" | "integer" | "float" | "boolean";
 
-// Dataframe field definition
-export interface DataframeField {
-  name: string;
-  type: FieldType;
-}
-
-// Dataframe schema definition
-export interface DataframeSchema {
-  fields: DataframeField[];
-  primaryKey: string[];
-}
-
-// Tabular dataframe representation
-export interface DataframeData {
-  schema: DataframeSchema;
-  data: Record<string, any>[];
-}
-
 // Import analysis status for individual documents (maps to ImportStatus enum in backend)
 export type ImportStatus = "add" | "update" | "skip" | "ignore" | "failed";
