+++ extralit-frontend/v1/domain/services/FileMatchingService.spec.js
@@ -1,4 +1,4 @@
-import { PdfMatchingService } from "./PdfMatchingService";
+import { PdfMatchingService } from "./FileMatchingService";
 
 describe("PdfMatchingService", () => {
   let service;
