+++ extralit-frontend/v1/domain/services/FileMatchingService.ts
@@ -1,4 +1,4 @@
-import type { IPdfMatchingService, PdfMatchResult, PdfMatchingResult } from "./IPdfMatchingService";
+import type { IFileMatchingService, FileMatchResult, FileMatchingResult } from "./IFileMatchingService";
 import type { ParsedEntry } from "./IFileService";
 
 interface MatchCandidate {
@@ -13,8 +13,8 @@ interface PrefixMatchResult {
   type: string;
 }
 
-export class PdfMatchingService implements IPdfMatchingService {
-  matchFiles(files: File[], entries: ParsedEntry[]): PdfMatchingResult {
+export class PdfMatchingService implements IFileMatchingService {
+  matchFiles(files: File[], entries: ParsedEntry[]): FileMatchingResult {
     if (!entries || entries.length === 0 || files.length === 0) {
       return {
         matchedFiles: [],
@@ -22,7 +22,7 @@ export class PdfMatchingService implements IPdfMatchingService {
       };
     }
 
-    const matchedFiles: PdfMatchResult[] = [];
+    const matchedFiles: FileMatchResult[] = [];
     const unmatchedFiles: File[] = [];
 
     // Track which references have been matched to support multiple files per reference
