+++ extralit-frontend/v1/domain/services/IFileMatchingService.ts
@@ -1,0 +1,17 @@
+import type { ParsedEntry } from "./IFileService";
+
+export interface FileMatchResult {
+  file: File;
+  bibEntry: ParsedEntry;
+  matchType: string;
+  confidence: number;
+}
+
+export interface FileMatchingResult {
+  matchedFiles: FileMatchResult[];
+  unmatchedFiles: File[];
+}
+
+export interface IFileMatchingService {
+  matchFiles(files: File[], entries: ParsedEntry[]): FileMatchingResult;
+}
