--- extralit-frontend/v1/domain/services/IPdfMatchingService.ts
@@ -1,17 +1,0 @@
-import type { ParsedEntry } from "./IFileService";
-
-export interface PdfMatchResult {
-  file: File;
-  bibEntry: ParsedEntry;
-  matchType: string;
-  confidence: number;
-}
-
-export interface PdfMatchingResult {
-  matchedFiles: PdfMatchResult[];
-  unmatchedFiles: File[];
-}
-
-export interface IPdfMatchingService {
-  matchFiles(files: File[], entries: ParsedEntry[]): PdfMatchingResult;
-}
